✨ feat(config): Consolidate world generation configurations
Moves gas pocket generation settings from MachineryWorldGenConfig to WorldGenConfig.
Renames MultiblockWorldGenConfig to WorldGenConfig for better clarity.
Updates references and language files accordingly.
