+++ src/main/resources/assets/omoshiroikamo/lang/en_US.lang
@@ -964,10 +964,13 @@ config.beaconRange=Range Settings
 config.solarArrayConfig=Solar Array Settings
 config.solarTierEnergy=Peak Energy per Tier
 config.solarCellSettings=Cell Settings
-config.eTWorldGenConfig=World Generation Settings
+config.worldGenConfig=World Generation Settings
 config.hardenedStoneGen=Hardened Stone
 config.alabasterGen=Alabaster
 config.basaltGen=Basalt
+config.heliumGasGen=Helium
+config.chlorineGasGen=Chlorine
+config.fluorineGasGen=Fluorine
 
 
 # NEI
