+++ src/main/resources/assets/omoshiroikamo/lang/ja_JP.lang
@@ -965,7 +965,7 @@ config.beaconRange=効果範囲
 config.solarArrayConfig=ソーラーアレイの設定
 config.solarTierEnergy=Tier別ピーク出力
 config.solarCellSettings=セル設定
-config.eTWorldGenConfig=ワールド生成の設定
+config.worldGenConfig=ワールド生成の設定
 config.hardenedStoneGen=硬化石
 config.alabasterGen=アラバスター
 config.basaltGen=玄武岩
